fix things

+++ src/libc/math.hpp
@@ -1,3 +1,3 @@
-static inline float floor(float n) { 
-    return (float) ((int) n - (n < 0.0f));
+static inline int floor(double n) { 
+    return (int) ((int) n - (n < 0.0f));
 }
